add host_utils and cleanup host_codegen
